Extract getting DisplayAccounts into use case
